refactor(p2): extract ModelCouncil into sibling page (stage 2 of 4)
SettingExperimentalModelCouncilPage + 3 private helpers (ModelCouncilSeatEditor,
ModelCouncilPropertyRow, ModelCouncilSelectRow<T>) moved from
SettingExperimentalPage.kt (post-stage1 L171-674) into new sibling
SettingExperimentalModelCouncilPage.kt.

Byte-equal: bodies and signatures unchanged.
Visibility unchanged (public page composable, private helpers).

Coordinator strip: 1920 -> 1398 (-522 with import cleanup).
Imports trimmed in coordinator: 18 orphans.
- 11 ModelCouncil-specific (DEFAULT_MODEL_COUNCIL_*, EXTENDED_MODEL_COUNCIL_*,
  MODEL_COUNCIL_EXTERNAL_MODEL_PLACEHOLDER, ModelCouncilRolePresets,
  ModelCouncilRuntimeSetting, ModelCouncilSeat, ModelCouncilSeatRunner)
- 6 shared symbols whose only remaining uses were inside this extraction
  (AgentPromptConfigRepository, ReasoningLevel, Uuid, ModelSelector,
  ModelType, ProviderSetting)
- 1 catch-up from stage 1 (findModelById was a SubAgentBuiltInRow-only
  dependency; missed during stage 1 cleanup, removed now)

Baseline 6/463 unit tests preserved.
